Update AUTHORS
This repo uses the terms "The Hush developers" and "The Supernet developers" without
them being defined, so we define them explicitly.

This repository was forked from jl777/komodo.git at a time when it had no LICENSE, AUTHORS
or DEVELOPER_AGREEMENT files. I advised jl777, as an expert with 23 years of experience
with open source and free software license, for him to fix it.

He did not listen to my advice correctly and hastily added some files to the LEGAL/ subdirectory
and added himself as sole author, without consulting all people who had commit rights in the
past and present. Additionally the new developer agreement including wording about closed
source binaries that I would never have agreed to:

https://github.com/jl777/komodo/tree/master/LEGAL

Every source code file in the repo refers to files in the root directory of the repo, so jl777's
"LEGAL hack" is not even legally binding in his own code repository, and doesn't even exist
in this repository, because Hush v3 forked from KMD before I advised jl777 to fix the issue in his repo.

For the purposes of this code repository, "The Supernet developers" refers to people who were in
AUTHORS previous to this commit. This commit clarifies the situation by defining "The Hush developers"
explicitly and the fact that jl777 should not be considered a Hush Developer, unless he agrees to our
new developer agreement.

Additionally, The Supernet developers only released software under MIT/GPLv2 and the new policy of
The Hush developers is to release all code under GPLv3 and to pursue violation enforcement to the
fullest extent possible by law.